Add plugin list in settings page.
Plugins that were installed via Galore are shown in the
collapsible section "Installed".
Once a plugin from that list is missing, it will be moved to
the "History" section (also collapsible), and marked with a timestamp.
History entries can either be removed individually or the whole list
can be cleared.